Move trust v1 handlers into their own file
