+++ libsplinter/src/network/auth/handlers/v1_handlers/mod.rs
@@ -14,6 +14,9 @@
 
 //! Message handlers for v1 authorization messages
 
+#[cfg(feature = "trust-authorization")]
+pub mod trust;
+
 #[cfg(feature = "challenge-authorization")]
 use cylinder::{PublicKey, Signature, Signer, Verifier};
 
@@ -23,9 +26,7 @@ use crate::network::auth::state_machine::challenge_v1::{
     ChallengeAuthorizationRemoteAction, ChallengeAuthorizationRemoteState,
 };
 #[cfg(feature = "trust-authorization")]
-use crate::network::auth::state_machine::trust_v1::{
-    TrustAuthorizationLocalAction, TrustAuthorizationRemoteAction, TrustAuthorizationRemoteState,
-};
+use crate::network::auth::state_machine::trust_v1::TrustAuthorizationLocalAction;
 use crate::network::auth::{
     AuthorizationLocalAction, AuthorizationLocalState, AuthorizationManagerStateMachine,
     AuthorizationMessage, AuthorizationRemoteAction, AuthorizationRemoteState,
@@ -34,6 +35,8 @@ use crate::network::auth::{
 use crate::network::dispatch::{
     ConnectionId, DispatchError, Handler, MessageContext, MessageSender,
 };
+#[cfg(feature = "trust-authorization")]
+use crate::protocol::authorization::AuthTrustRequest;
 #[cfg(feature = "challenge-authorization")]
 use crate::protocol::authorization::{
     AuthChallengeNonceRequest, AuthChallengeNonceResponse, AuthChallengeSubmitRequest,
@@ -43,8 +46,6 @@ use crate::protocol::authorization::{
     AuthComplete, AuthProtocolRequest, AuthProtocolResponse, AuthorizationError,
     PeerAuthorizationType,
 };
-#[cfg(feature = "trust-authorization")]
-use crate::protocol::authorization::{AuthTrustRequest, AuthTrustResponse};
 use crate::protocol::network::NetworkMessage;
 use crate::protocol::{PEER_AUTHORIZATION_PROTOCOL_MIN, PEER_AUTHORIZATION_PROTOCOL_VERSION};
 use crate::protos::authorization;
@@ -478,181 +479,6 @@ impl Handler for AuthProtocolResponseHandler {
             }
             Ok(next_state) => panic!("Should not have been able to transition to {}", next_state),
         }
-        Ok(())
-    }
-}
-
-/// Handler for the Authorization Trust Request Message Type
-#[cfg(feature = "trust-authorization")]
-pub struct AuthTrustRequestHandler {
-    auth_manager: AuthorizationManagerStateMachine,
-}
-
-#[cfg(feature = "trust-authorization")]
-impl AuthTrustRequestHandler {
-    pub fn new(auth_manager: AuthorizationManagerStateMachine) -> Self {
-        Self { auth_manager }
-    }
-}
-
-#[cfg(feature = "trust-authorization")]
-impl Handler for AuthTrustRequestHandler {
-    type Source = ConnectionId;
-    type MessageType = authorization::AuthorizationMessageType;
-    type Message = authorization::AuthTrustRequest;
-
-    fn match_type(&self) -> Self::MessageType {
-        authorization::AuthorizationMessageType::AUTH_TRUST_REQUEST
-    }
-
-    fn handle(
-        &self,
-        msg: Self::Message,
-        context: &MessageContext<Self::Source, Self::MessageType>,
-        sender: &dyn MessageSender<Self::Source>,
-    ) -> Result<(), DispatchError> {
-        debug!(
-            "Received authorization trust request from {}",
-            context.source_connection_id()
-        );
-        let trust_request = AuthTrustRequest::from_proto(msg)?;
-        match self.auth_manager.next_remote_state(
-            context.source_connection_id(),
-            AuthorizationRemoteAction::Trust(
-                TrustAuthorizationRemoteAction::ReceiveAuthTrustRequest(Identity::Trust {
-                    identity: trust_request.identity.to_string(),
-                }),
-            ),
-        ) {
-            Err(err) => {
-                send_authorization_error(
-                    &self.auth_manager,
-                    context.source_id(),
-                    context.source_connection_id(),
-                    sender,
-                    &err.to_string(),
-                )?;
-                return Ok(());
-            }
-            Ok(AuthorizationRemoteState::Trust(
-                TrustAuthorizationRemoteState::ReceivedAuthTrustRequest(_),
-            )) => {
-                debug!(
-                    "Sending trust response to connection {} after receiving identity {}",
-                    context.source_connection_id(),
-                    trust_request.identity,
-                );
-                let auth_msg = AuthorizationMessage::AuthTrustResponse(AuthTrustResponse);
-                let msg_bytes = IntoBytes::<network::NetworkMessage>::into_bytes(
-                    NetworkMessage::from(auth_msg),
-                )?;
-                sender
-                    .send(context.source_id().clone(), msg_bytes)
-                    .map_err(|(recipient, payload)| {
-                        DispatchError::NetworkSendError((recipient.into(), payload))
-                    })?;
-            }
-            Ok(next_state) => panic!("Should not have been able to transition to {}", next_state),
-        }
-
-        if self
-            .auth_manager
-            .next_remote_state(
-                context.source_connection_id(),
-                AuthorizationRemoteAction::Trust(
-                    TrustAuthorizationRemoteAction::SendAuthTrustResponse,
-                ),
-            )
-            .is_err()
-        {
-            error!("Unable to transition from ReceivedAuthTrustRequest to Done")
-        };
-
-        Ok(())
-    }
-}
-
-#[cfg(feature = "trust-authorization")]
-/// Handler for the Authorization Trust Response Message Type
-pub struct AuthTrustResponseHandler {
-    auth_manager: AuthorizationManagerStateMachine,
-}
-
-#[cfg(feature = "trust-authorization")]
-impl AuthTrustResponseHandler {
-    pub fn new(auth_manager: AuthorizationManagerStateMachine) -> Self {
-        Self { auth_manager }
-    }
-}
-
-#[cfg(feature = "trust-authorization")]
-impl Handler for AuthTrustResponseHandler {
-    type Source = ConnectionId;
-    type MessageType = authorization::AuthorizationMessageType;
-    type Message = authorization::AuthTrustResponse;
-
-    fn match_type(&self) -> Self::MessageType {
-        authorization::AuthorizationMessageType::AUTH_TRUST_RESPONSE
-    }
-
-    fn handle(
-        &self,
-        _msg: Self::Message,
-        context: &MessageContext<Self::Source, Self::MessageType>,
-        sender: &dyn MessageSender<Self::Source>,
-    ) -> Result<(), DispatchError> {
-        debug!(
-            "Received authorization trust response from {}",
-            context.source_connection_id()
-        );
-        match self.auth_manager.next_local_state(
-            context.source_connection_id(),
-            AuthorizationLocalAction::Trust(
-                TrustAuthorizationLocalAction::ReceiveAuthTrustResponse,
-            ),
-        ) {
-            Err(err) => {
-                send_authorization_error(
-                    &self.auth_manager,
-                    context.source_id(),
-                    context.source_connection_id(),
-                    sender,
-                    &err.to_string(),
-                )?;
-                return Ok(());
-            }
-            Ok(AuthorizationLocalState::Authorized) => (),
-            Ok(next_state) => panic!("Should not have been able to transition to {}", next_state),
-        }
-
-        let auth_msg = AuthorizationMessage::AuthComplete(AuthComplete);
-        let msg_bytes =
-            IntoBytes::<network::NetworkMessage>::into_bytes(NetworkMessage::from(auth_msg))?;
-        sender
-            .send(context.source_id().clone(), msg_bytes)
-            .map_err(|(recipient, payload)| {
-                DispatchError::NetworkSendError((recipient.into(), payload))
-            })?;
-
-        match self.auth_manager.next_local_state(
-            context.source_connection_id(),
-            AuthorizationLocalAction::SendAuthComplete,
-        ) {
-            Err(err) => {
-                send_authorization_error(
-                    &self.auth_manager,
-                    context.source_id(),
-                    context.source_connection_id(),
-                    sender,
-                    &err.to_string(),
-                )?;
-                return Ok(());
-            }
-            Ok(AuthorizationLocalState::WaitForComplete) => (),
-            Ok(AuthorizationLocalState::AuthorizedAndComplete) => (),
-            Ok(next_state) => panic!("Should not have been able to transition to {}", next_state),
-        };
-
         Ok(())
     }
 }
@@ -1229,14 +1055,13 @@ mod tests {
     use std::sync::{Arc, Mutex};
 
     use cylinder::{
-        secp256k1::Secp256k1Context, Context, PublicKey, Signature, VerificationError, Verifier,
+        secp256k1::Secp256k1Context, Context, PublicKey, Signature, Signer, VerificationError,
+        Verifier,
     };
     use protobuf::Message;
 
     #[cfg(feature = "challenge-authorization")]
     use crate::network::auth::state_machine::challenge_v1::ChallengeAuthorizationLocalState;
-    #[cfg(feature = "trust-authorization")]
-    use crate::network::auth::state_machine::trust_v1::TrustAuthorizationLocalState;
     use crate::network::auth::AuthorizationDispatchBuilder;
     use crate::network::auth::ManagedAuthorizationState;
     use crate::protos::network::NetworkMessageType;
@@ -1392,384 +1217,6 @@ mod tests {
                 .get_accepted_authorization_type()
                 .to_vec()
         );
-    }
-
-    /// Test that an auth protocol response is properly handled via the dispatcher when trust is
-    /// is set as the accepted authorization type.
-    ///
-    /// This is verified by:
-    ///
-    /// 1) no error from the dispatcher
-    /// 2) the handler should send an AuthTrustRequest with the provided identity
-    #[test]
-    #[cfg(feature = "trust-authorization")]
-    fn protocol_response_trust() {
-        let connection_id = "test_connection".to_string();
-        // need to setup expected authorization state
-        let auth_mgr = AuthorizationManagerStateMachine::default();
-        auth_mgr
-            .shared
-            .lock()
-            .expect("lock poisoned")
-            .states
-            .insert(
-                connection_id.to_string(),
-                ManagedAuthorizationState {
-                    local_state: AuthorizationLocalState::WaitingForAuthProtocolResponse,
-                    remote_state: AuthorizationRemoteState::SentAuthProtocolResponse,
-                    received_complete: false,
-                },
-            );
-        let mock_sender = MockSender::new();
-        let dispatch_sender = mock_sender.clone();
-        // mut is required if chalenge authorization is enabled
-        #[allow(unused_mut)]
-        let mut dispatcher_builder =
-            AuthorizationDispatchBuilder::new().with_identity("mock_identity");
-
-        #[cfg(feature = "challenge-authorization")]
-        {
-            dispatcher_builder = dispatcher_builder
-                .with_signers(&[new_signer()])
-                .with_nonce(&vec![])
-                .with_expected_authorization(None)
-                .with_local_authorization(None)
-                .with_verifier(Box::new(NoopVerifier))
-        }
-
-        let dispatcher = dispatcher_builder
-            .build(dispatch_sender, auth_mgr)
-            .expect("Unable to build authorization dispatcher");
-
-        let msg_bytes = IntoBytes::<authorization::AuthorizationMessage>::into_bytes(
-            AuthorizationMessage::AuthProtocolResponse(AuthProtocolResponse {
-                auth_protocol: 1,
-                accepted_authorization_type: vec![PeerAuthorizationType::Trust],
-            }),
-        )
-        .expect("Unable to get message bytes");
-
-        assert_eq!(
-            Ok(()),
-            dispatcher.dispatch(
-                connection_id.clone().into(),
-                &NetworkMessageType::AUTHORIZATION,
-                msg_bytes
-            )
-        );
-
-        let (recipient, message_bytes) = mock_sender
-            .next_outbound()
-            .expect("Unable to receive message over the network");
-        let recipient: String = recipient.into();
-        assert_eq!(&connection_id, &recipient);
-
-        let trust_request: authorization::AuthTrustRequest = expect_auth_message(
-            authorization::AuthorizationMessageType::AUTH_TRUST_REQUEST,
-            &message_bytes,
-        );
-        assert_eq!("mock_identity", trust_request.get_identity());
-    }
-
-    /// Test that a trust request is properly handled. Also verify end state is set to
-    /// WaitingForAuthTrustResponse and Done.
-    ///
-    /// This is verified by:
-    ///
-    /// 1) no error from the dispatcher
-    /// 2) the handler should send AuthTrustResponse Message
-    /// 3) verify the states are set to WaitingForAuthTrustResponse and Done(identity)
-    #[test]
-    #[cfg(feature = "trust-authorization")]
-    fn trust_request() {
-        let connection_id = "test_connection".to_string();
-        // need to setup expected authorization state
-        let auth_mgr = AuthorizationManagerStateMachine::default();
-        auth_mgr
-            .shared
-            .lock()
-            .expect("lock poisoned")
-            .states
-            .insert(
-                connection_id.to_string(),
-                ManagedAuthorizationState {
-                    local_state: AuthorizationLocalState::Trust(
-                        TrustAuthorizationLocalState::WaitingForAuthTrustResponse,
-                    ),
-                    remote_state: AuthorizationRemoteState::SentAuthProtocolResponse,
-                    received_complete: false,
-                },
-            );
-        let mock_sender = MockSender::new();
-        let dispatch_sender = mock_sender.clone();
-        // mut is required if chalenge authorization is enabled
-        #[allow(unused_mut)]
-        let mut dispatcher_builder =
-            AuthorizationDispatchBuilder::new().with_identity("mock_identity");
-
-        #[cfg(feature = "challenge-authorization")]
-        {
-            dispatcher_builder = dispatcher_builder
-                .with_signers(&[new_signer()])
-                .with_nonce(&vec![])
-                .with_expected_authorization(None)
-                .with_local_authorization(None)
-                .with_verifier(Box::new(NoopVerifier))
-        }
-
-        let dispatcher = dispatcher_builder
-            .build(dispatch_sender, auth_mgr.clone())
-            .expect("Unable to build authorization dispatcher");
-
-        let msg_bytes = IntoBytes::<authorization::AuthorizationMessage>::into_bytes(
-            AuthorizationMessage::AuthTrustRequest(AuthTrustRequest {
-                identity: "other_identity".to_string(),
-            }),
-        )
-        .expect("Unable to get message bytes");
-
-        assert_eq!(
-            Ok(()),
-            dispatcher.dispatch(
-                connection_id.clone().into(),
-                &NetworkMessageType::AUTHORIZATION,
-                msg_bytes
-            )
-        );
-
-        let (recipient, message_bytes) = mock_sender
-            .next_outbound()
-            .expect("Unable to receive message over the network");
-        let recipient: String = recipient.into();
-        assert_eq!(&connection_id, &recipient);
-
-        let _trust_response: authorization::AuthTrustResponse = expect_auth_message(
-            authorization::AuthorizationMessageType::AUTH_TRUST_RESPONSE,
-            &message_bytes,
-        );
-
-        let managed_state = auth_mgr
-            .shared
-            .lock()
-            .expect("lock poisoned")
-            .states
-            .get(&connection_id)
-            .cloned()
-            .expect("missing managed state for connection id");
-
-        assert_eq!(
-            managed_state.local_state,
-            AuthorizationLocalState::Trust(
-                TrustAuthorizationLocalState::WaitingForAuthTrustResponse
-            )
-        );
-        assert_eq!(
-            managed_state.remote_state,
-            AuthorizationRemoteState::Done(Identity::Trust {
-                identity: "other_identity".to_string()
-            })
-        );
-        assert_eq!(managed_state.received_complete, false);
-    }
-
-    /// Test that a trust response is properly handled. Also verify end state is set to
-    /// WaitForComplete because received_complete is set to false
-    ///
-    /// This is verified by:
-    ///
-    /// 1) no error from the dispatcher
-    /// 2) the handler should send AuthComplete Message
-    /// 3) verify that because auth complete has not been received, the states are set to
-    ///    WaitingForComplete and Done(identity)
-    #[test]
-    #[cfg(feature = "trust-authorization")]
-    fn trust_response() {
-        let connection_id = "test_connection".to_string();
-        // need to setup expected authorization state
-        let auth_mgr = AuthorizationManagerStateMachine::default();
-        auth_mgr
-            .shared
-            .lock()
-            .expect("lock poisoned")
-            .states
-            .insert(
-                connection_id.to_string(),
-                ManagedAuthorizationState {
-                    local_state: AuthorizationLocalState::Trust(
-                        TrustAuthorizationLocalState::WaitingForAuthTrustResponse,
-                    ),
-                    remote_state: AuthorizationRemoteState::Done(Identity::Trust {
-                        identity: "other_identity".to_string(),
-                    }),
-                    received_complete: false,
-                },
-            );
-        let mock_sender = MockSender::new();
-        let dispatch_sender = mock_sender.clone();
-        // mut is required if chalenge authorization is enabled
-        #[allow(unused_mut)]
-        let mut dispatcher_builder =
-            AuthorizationDispatchBuilder::new().with_identity("mock_identity");
-
-        #[cfg(feature = "challenge-authorization")]
-        {
-            dispatcher_builder = dispatcher_builder
-                .with_signers(&[new_signer()])
-                .with_nonce(&vec![])
-                .with_expected_authorization(None)
-                .with_local_authorization(None)
-                .with_verifier(Box::new(NoopVerifier))
-        }
-
-        let dispatcher = dispatcher_builder
-            .build(dispatch_sender, auth_mgr.clone())
-            .expect("Unable to build authorization dispatcher");
-
-        let msg_bytes = IntoBytes::<authorization::AuthorizationMessage>::into_bytes(
-            AuthorizationMessage::AuthTrustResponse(AuthTrustResponse),
-        )
-        .expect("Unable to get message bytes");
-
-        assert_eq!(
-            Ok(()),
-            dispatcher.dispatch(
-                connection_id.clone().into(),
-                &NetworkMessageType::AUTHORIZATION,
-                msg_bytes
-            )
-        );
-
-        let (recipient, message_bytes) = mock_sender
-            .next_outbound()
-            .expect("Unable to receive message over the network");
-        let recipient: String = recipient.into();
-        assert_eq!(&connection_id, &recipient);
-
-        let _trust_response: authorization::AuthComplete = expect_auth_message(
-            authorization::AuthorizationMessageType::AUTH_COMPLETE,
-            &message_bytes,
-        );
-
-        let managed_state = auth_mgr
-            .shared
-            .lock()
-            .expect("lock poisoned")
-            .states
-            .get(&connection_id)
-            .cloned()
-            .expect("missing managed state for connection id");
-
-        assert_eq!(
-            managed_state.local_state,
-            AuthorizationLocalState::WaitForComplete,
-        );
-        assert_eq!(
-            managed_state.remote_state,
-            AuthorizationRemoteState::Done(Identity::Trust {
-                identity: "other_identity".to_string()
-            })
-        );
-        assert_eq!(managed_state.received_complete, false);
-    }
-
-    /// Test that a trust response is properly handled. Also verify end state is set to
-    /// AuthorizedAndComplete because received_complete is set to true
-    ///
-    /// This is verified by:
-    ///
-    /// 1) no error from the dispatcher
-    /// 2) the handler should send AuthComplete Message
-    /// 3) verify that because auth complete has  been received, the states are set to
-    ///    AuthorizedAndComplete and Done(identity)
-    #[test]
-    #[cfg(feature = "trust-authorization")]
-    fn trust_response_complete() {
-        let connection_id = "test_connection".to_string();
-        // need to setup expected authorization state
-        let auth_mgr = AuthorizationManagerStateMachine::default();
-        auth_mgr
-            .shared
-            .lock()
-            .expect("lock poisoned")
-            .states
-            .insert(
-                connection_id.to_string(),
-                ManagedAuthorizationState {
-                    local_state: AuthorizationLocalState::Trust(
-                        TrustAuthorizationLocalState::WaitingForAuthTrustResponse,
-                    ),
-                    remote_state: AuthorizationRemoteState::Done(Identity::Trust {
-                        identity: "other_identity".to_string(),
-                    }),
-                    received_complete: true,
-                },
-            );
-        let mock_sender = MockSender::new();
-        let dispatch_sender = mock_sender.clone();
-        // mut is required if chalenge authorization is enabled
-        #[allow(unused_mut)]
-        let mut dispatcher_builder =
-            AuthorizationDispatchBuilder::new().with_identity("mock_identity");
-
-        #[cfg(feature = "challenge-authorization")]
-        {
-            dispatcher_builder = dispatcher_builder
-                .with_signers(&[new_signer()])
-                .with_nonce(&vec![])
-                .with_expected_authorization(None)
-                .with_local_authorization(None)
-                .with_verifier(Box::new(NoopVerifier))
-        }
-
-        let dispatcher = dispatcher_builder
-            .build(dispatch_sender, auth_mgr.clone())
-            .expect("Unable to build authorization dispatcher");
-
-        let msg_bytes = IntoBytes::<authorization::AuthorizationMessage>::into_bytes(
-            AuthorizationMessage::AuthTrustResponse(AuthTrustResponse),
-        )
-        .expect("Unable to get message bytes");
-
-        assert_eq!(
-            Ok(()),
-            dispatcher.dispatch(
-                connection_id.clone().into(),
-                &NetworkMessageType::AUTHORIZATION,
-                msg_bytes
-            )
-        );
-
-        let (recipient, message_bytes) = mock_sender
-            .next_outbound()
-            .expect("Unable to receive message over the network");
-        let recipient: String = recipient.into();
-        assert_eq!(&connection_id, &recipient);
-
-        let _trust_response: authorization::AuthComplete = expect_auth_message(
-            authorization::AuthorizationMessageType::AUTH_COMPLETE,
-            &message_bytes,
-        );
-
-        let managed_state = auth_mgr
-            .shared
-            .lock()
-            .expect("lock poisoned")
-            .states
-            .get(&connection_id)
-            .cloned()
-            .expect("missing managed state for connection id");
-
-        assert_eq!(
-            managed_state.local_state,
-            AuthorizationLocalState::AuthorizedAndComplete,
-        );
-        assert_eq!(
-            managed_state.remote_state,
-            AuthorizationRemoteState::Done(Identity::Trust {
-                identity: "other_identity".to_string()
-            })
-        );
-        assert_eq!(managed_state.received_complete, true);
     }
 
     /// Test that a protocol response is properly handled when only challenge is in
